plant_api/requirements.txt
fastapi
uvicorn[standard]
onnxruntime
pillow
numpy